requirements.txt
uvicorn
pyinstaller
google-genai
subprocces
python-dotenv
